Make use of the 'Fault bit' in the 'RTD Resistance Registers'
Make use of the 'Fault bit' in the 'RTD Resistance Registers' to save some SPI-traffic.
When reading the  'RTD Resistance Registers' we get the information if it makes sense to read the 'Fault Status Register' for free. We don't have to read the register over SPI if the register is empty anyway.