Normalize resistance before calculating positive temperature
Normalize resistance before calculating temperature with the formula for the positive temperatures.
Positive temperature means (when normalized) resistance > 100 Ohm.
We then can decide what formula to use, before we calculated the positive one.

This should significantly speed up the calculation for negative temperatures.

Even more efficient would be to store the normalized value for Rref in the object - but that would change the API.